producer/requirements.txt
Mastodon.py==1.8.1
kafka-python==2.0.2
python-dotenv==1.0.1
confluent-kafka==2.5.0
